Change: (lint-indent) Use custom function
The package indent-lint is designed to be able to run asynchronously
in an interactive Emacs session, and through its dependencies it
brings in 4 packages to do so.  Our batch-and-exit needs are very
simple, so this function suffices and requires no dependencies.  It
also uses Emacs's undo facility to detect which lines have changed,
which doesn't require calling out to "diff".