fix color escaping
Linux man page[1] contains a list of console codes and among others it says that
`ESC` (`0x1B`, `^[`) starts an escape sequence. `0x1B` (hex) is `033` (oct).

`echo` command on Linux also treats[3] `\e` as escape. But on BSD `\e` is not
treated[2] as escape.

This also means that on BSD derived systems (including macOS) the output of
makem.sh is printed in an ugly fashion.

I think it makes sense to improve portability of makem.sh by using `\033`.

  [1]: https://www.unix.com/man-page/linux/4/console_codes/
  [2]: https://www.unix.com/man-page/bsd/1/echo/
  [3]: https://www.unix.com/man-page/linux/1/echo/