Fix: (lint-buttercup) Don't use buttercup-run-discover
The buttercup-run-discover command searches the current directory
recursively and loads all files with a test-like name in them.  This
can cause a problem when a sandbox directory exists in the current
directory which includes a test-like file.  For example, the relint
package from ELPA includes its relint-test.el file in the package, which
buttercup-run-discover finds and loads, which causes it to try to load
relint, even though relint isn't needed.

This is especially confusing when makem.sh is not running with
--sandbox, because makem.sh is running the local ~/.emacs.d
configuration, but buttercup-run-discover is causing files from the
sandbox to be loaded, which may try to load files that are not present
in the ~/.emacs.d config.

But we don't need Buttercup to discover its own tests anyway, because
by the time we run Emacs, we have already ensured that Buttercup tests
are present.  So we can just load the test files directly and then
call buttercup-run, which runs already-loaded Buttercup tests.

This should also improve performance, especially when there's a
sandbox with a lot of files that buttercup-run-discover would check
the names of.